+++ smartcosmos-core/src/main/java/net/smartcosmos/objects/constraints/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Custom constraint validation.
+ */
+package net.smartcosmos.objects.constraints;