+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add separate SDKs for different AWS services to `smartcosmos-dependencies` (SCE-19)
 * Fix smartcosmos dependency scope in `smartcosmos-extension-archetype`
 * Extract Geospatial from Objects into separate "Geospatial Extension" (OBJECTS-622)
+* OBJECTS-377: Add data validation annotation for Metadata
 
 ## Release 2.13.1 (March 8, 2016)